+++ src/test/java/org/jolokia/docker/maven/config/handler/property/PropertyConfigHandlerTest.java
@@ -242,6 +242,8 @@ public class PropertyConfigHandlerTest {
         WaitConfiguration wait = runConfig.getWaitConfiguration();
         assertEquals("http://foo.com", wait.getUrl());
         assertEquals("pattern", wait.getLog());
+        assertEquals("post_start_command", wait.getExec().getPostStart());
+        assertEquals("pre_stop_command", wait.getExec().getPreStop());
         assertEquals(5, wait.getTime());
     }
 
@@ -301,6 +303,8 @@ public class PropertyConfigHandlerTest {
                 k(USER), "tomcat",
                 k(VOLUMES) + ".1","/foo",
                 k(VOLUMES_FROM) + ".1", "from",
+                k(PRE_STOP), "pre_stop_command",
+                k(POST_START), "post_start_command",
                 k(WAIT_LOG), "pattern",
                 k(WAIT_TIME), "5",
                 k(WAIT_URL), "http://foo.com",